refactor(action): fold RestoreSnapshotIos into RestoreSnapshot platform switch
Merge the Android and iOS snapshot-restore commands into a single
RestoreSnapshot class that dispatches on device.platform via a switch,
with private executeAndroid/executeIos methods and Ios-prefixed iOS
helpers (issue #5461). Replaces the platform branch in
deviceSnapshotManager.createRestoreProvider with a single construction;
the Android-only constructor guard is dropped in favor of the switch.
Updates the SnapshotProvider interface doc comments for both folded
capture and restore classes.

Behavior-preserving. iOS coverage moves into RestoreSnapshot.test.ts and
the SnapshotProvider conformance case targets the folded class. The iOS
app-data restore path is split into resolveIosAppDataPath and
restoreIosBundleContainer helpers to keep the merged class within the
existing complexity/max-depth budgets (no baseline growth); stale
RestoreSnapshotIos baseline entries are pruned.
